refactor: remove mapUriToPath and move get handler into separate file
