+++ src/http-response/http-response.cpp
@@ -78,96 +78,6 @@ std::string HttpResponse::toString() {
     return responseStream.str();
 }
 
-// constructs the filesystem path given a docRoot and request path
-std::string mapUriToPath(const std::string& docRoot, const std::string& uri) {
-    std::string safeUri = uri;
-    if (safeUri.empty() || safeUri[0] != '/')
-        safeUri = "/" + safeUri;
-
-    // Normalize “/” -> “/index.html” (i.e. default file)
-    if (safeUri == "/")
-        safeUri = "/index.html";
-
-    std::cout << "Mapping URI: " << uri << " to path: " << docRoot + safeUri << "\n";
-    return docRoot + safeUri;
-}
-
-// Reads the entire file at filepath into a std::string.
-// Returns empty string if any error occurs.
-std::string readFileToString(const std::string& filepath) {
-    struct stat st;
-    if (stat(filepath.c_str(), &st) < 0 || !S_ISREG(st.st_mode)) {
-        return "";  // file doesn’t exist or isn’t regular
-    }
-
-    int fd = open(filepath.c_str(), O_RDONLY);
-    if (fd < 0)
-        return "";
-
-    std::string content;
-    content.reserve(st.st_size);
-
-    const size_t BUFFER_SIZE = 4096;
-    char buffer[BUFFER_SIZE];
-    ssize_t n;
-    while ((n = read(fd, buffer, BUFFER_SIZE)) > 0) {
-        content.append(buffer, n);
-    }
-    close(fd);
-    return content;
-}
-
-
-std::string getMimeType(const std::string& path)
-{
-    size_t dot = path.rfind('.');
-    if (dot == std::string::npos)
-        return "application/octet-stream";
-
-    std::string ext = path.substr(dot + 1);
-    if (ext == "html" || ext == "htm") return "text/html";
-    if (ext == "css")                  return "text/css";
-    if (ext == "js")                   return "application/javascript";
-    if (ext == "png")                  return "image/png";
-    if (ext == "jpg" || ext == "jpeg") return "image/jpeg";
-    if (ext == "gif")                  return "image/gif";
-    if (ext == "txt")                  return "text/plain";
-    if (ext == "json")                 return "text/plain";
-    if (ext == "cgi")                  return "text/plain";
-    if (ext == "py")                   return "text/plain";
-    // Fallback
-    return "application/octet-stream";
-}
-
-void HttpResponse::handleGetRequest(const std::string& uri, Server &serverConfig, const Location& location)
-{
-    std::string fullPath = resolveHttpPath(uri, serverConfig);
-
-    struct stat info;
-    if (stat(fullPath.c_str(), &info) < 0)
-        throw HttpException(NOT_FOUND, "Path is not a file or directory");
-
-    if (S_ISDIR(info.st_mode) && location.autoIndex) {
-        std::cout << GREEN "AutoIndex found\n" RESET; //// debug
-        std::string directoryContent = readDirectorytoString(fullPath, uri);
-        setStatus(OK);
-        setHeader("Content-Type", "text/html");
-        setBody(directoryContent);
-        return ;
-    }
-    else if (S_ISREG(info.st_mode))
-    {
-        std::cout << GREEN "Static file found\n" RESET; //// debug
-        std::string fileContents = readFileToString(fullPath);
-        if (fileContents.empty())
-            throw HttpException(NOT_FOUND, "File not found or readable");
-        std::string mime = getMimeType(fullPath);
-        setHeader("Content-Type", mime);
-        setBody(fileContents);
-        return ;
-    }
-}
-
 void HttpResponse::clearResponse()
 {
     _status = OK;
